Update and rename types.ts to types.ts
